Cleanup api folder
