+++ backend/api/tests/test_api.py
@@ -1,8 +1,8 @@
 from django.test import TestCase
 from django.urls import reverse
 
-from api import constants
 from categories.factories import CategoryFactory
+from core import constants
 from questions.factories import QuestionFactory
 from quizs.factories import QuizFactory
 from quizs.models import QuizQuestion
