+++ backend/core/templates/admin/base.html
