+++ backend/core/utils/utilities.py
