+++ backend/questions/admin.py
@@ -9,10 +9,11 @@ from import_export import fields, resources
 from import_export.admin import ImportMixin
 from import_export.widgets import ForeignKeyWidget, ManyToManyWidget
 
-from api import constants as api_constants, utilities_notion
 from categories.models import Category
+from core import constants as api_constants
 from core.admin import ExportMixin, admin_site
 from core.models import Configuration
+from core.utils import notion
 from questions.models import Question
 from quizs.models import Quiz
 from tags.models import Tag
@@ -84,7 +85,7 @@ class QuestionResource(resources.ModelResource):
     def after_import(self, dataset, result, using_transactions, dry_run, **kwargs):
         # result.totals: OrderedDict, keys: 'new', 'update', 'delete', 'skip', 'error', 'invalid'
         if not settings.DEBUG and not dry_run and not (result.totals["error"] or result.totals["invalid"]):  # noqa
-            utilities_notion.add_import_stats_row(result.total_rows, result.totals["new"], result.totals["update"])
+            notion.add_import_stats_row(result.total_rows, result.totals["new"], result.totals["update"])
         super(QuestionResource, self).after_import(dataset, result, using_transactions, dry_run, **kwargs)
 
     class Meta:
