+++ backend/stats/tests/test_command_generate_daily_stats.py
