+++ backend/stats/utilities.py
@@ -2,9 +2,9 @@ from datetime import date, datetime, timedelta
 
 from django.db.models import Count, F
 
-from api import constants
 from categories.models import Category
 from contributions.models import Contribution
+from core import constants
 from questions.models import Question
 from quizs.models import Quiz
 from stats.models import DailyStat, QuizAnswerEvent, QuizFeedbackEvent
